+++ src/command_modules/azure-cli-resource/HISTORY.rst
@@ -2,6 +2,7 @@
 
 Release History
 ===============
+* 'az rest': new command to sent out any REST request 
 
 2.1.15
 ++++++